tests/wpt/tests/css/css-align/parsing/grid-gap-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: grid-gap sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-align/#grid-gap-legacy">
<link rel="author" title="Takuya Kurimoto" href="mailto:takuya004869@gmail.com">
<meta name="assert" content="grid-gap supports the full grammar '<grid-row-gap> <column-gap>?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
test_shorthand_value('gap', 'normal', {
  'row-gap': 'normal',
  'column-gap': 'normal'
});

test_shorthand_value('gap', '10px 20%', {
  'row-gap': '10px',
  'column-gap': '20%'
});

test_shorthand_value('gap', '10px normal', {
  'row-gap': '10px',
  'column-gap': 'normal'
});

test_shorthand_value('gap', 'normal calc(20% + 10px)', {
  'grid-row-gap': 'normal',
  'grid-column-gap': 'calc(20% + 10px)'
});
</script>
</body>
</html>
